Add Geo restriction dashboard
