+++ configs/geo-rules.yaml
@@ -1,0 +1,29 @@
+rules:
+- action: block
+  country_name: null
+  created_at: 2026-06-30T16:24:31.611990379+00:00
+  enabled: true
+  id: 1
+  iso_code: CN
+  scope: global
+- action: block
+  country_name: null
+  created_at: 2026-06-30T16:24:31.632006212+00:00
+  enabled: true
+  id: 2
+  iso_code: RU
+  scope: global
+- action: block
+  country_name: null
+  created_at: 2026-06-30T16:24:31.651770004+00:00
+  enabled: true
+  id: 3
+  iso_code: KP
+  scope: global
+- action: challenge
+  country_name: null
+  created_at: 2026-06-30T16:24:31.669428421+00:00
+  enabled: true
+  id: 4
+  iso_code: IR
+  scope: global
